ENDPOINTS [#177] Refactor api requests to same structure as auth endpoints
