+++ containers/frontend/web/src/components/composites/user-item/user-item.styles.ts
@@ -1,6 +1,6 @@
 import { cn } from '@/lib/styles/cn';
 
-const userItemBase = 'w-full p-3';
+const userItemBase = 'w-full p-3 flex-wrap';
 
 export function userItemStyles({ className }: { className?: string } = {}) {
   return cn(userItemBase, className);
