+++ containers/frontend/web/src/features/social/actions/social.actions.ts
@@ -1,46 +1,6 @@
 import { fetchClient } from '@/lib/http/fetcher.client';
-import type {
-  GetFriendsListResponse,
-  GetPendingRequestsResponse,
-  GetSentRequestsResponse,
-  RespondFriendRequestResponse,
-} from '@/contracts/api/friendships/friendships.contracts';
 
-/**
- * Fetches the current user's accepted friends list.
- */
-export async function getFriendsList() {
-  const response = await fetchClient<GetFriendsListResponse>('/api/friends', 'GET', undefined, {
-    withAuth: true,
-  });
-  return response.data;
-}
-
-/**
- * Fetches pending incoming friend requests.
- */
-export async function getPendingRequests() {
-  const response = await fetchClient<GetPendingRequestsResponse>(
-    '/api/friends/requests/pending',
-    'GET',
-    undefined,
-    { withAuth: true },
-  );
-  return response.data;
-}
-
-/**
- * Fetches friend requests sent by the current user.
- */
-export async function getSentRequests() {
-  const response = await fetchClient<GetSentRequestsResponse>(
-    '/api/friends/requests/sent',
-    'GET',
-    undefined,
-    { withAuth: true },
-  );
-  return response.data;
-}
+import type { RespondFriendRequestResponse } from '@/contracts/api/friendships/friendships.contracts';
 
 /**
  * Accepts or rejects a friend request.
@@ -49,8 +9,13 @@ export async function respondToRequest(friendshipId: string, action: 'accept' | 
   const response = await fetchClient<RespondFriendRequestResponse>(
     '/api/friends/respond',
     'POST',
-    { friendshipId, action },
-    { withAuth: true },
+    {
+      friendshipId,
+      action,
+    },
+    {
+      withAuth: true,
+    },
   );
   return response.data;
 }
